# mandates/openapi.yaml
openapi: 3.1.0
info:
  title: Urban Heat Island Analysis API
  version: 1.0.0
  description: |
    Backend API for the Urban Heat Island Analysis Platform.
    All heavy computation runs via Google Earth Engine.
    Pixel data never passes through this API — only tile URLs and statistical summaries.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://uhi-api.onrender.com
    description: Production (Render)

# ─────────────────────────────────────────────────────────────────────────────
# SHARED SCHEMAS
# ─────────────────────────────────────────────────────────────────────────────
components:
  schemas:
    BBox:
      type: array
      description: Bounding box in WGS84. Format [minLon, minLat, maxLon, maxLat].
      items:
        type: number
        format: float
      minItems: 4
      maxItems: 4
      example: [72.77, 18.89, 72.99, 19.27]
    DateString:
      type: string
      format: date
      description: ISO 8601 date (YYYY-MM-DD). Must not be before 2001-01-01 (MODIS launch).
      example: "2023-06-01"
    TemperatureStats:
      type: object
      properties:
        min_temp:
          type: number
          format: float
          description: Minimum LST in the region (°C)
        max_temp:
          type: number
          format: float
          description: Maximum LST in the region (°C)
        mean_temp:
          type: number
          format: float
          description: Mean LST in the region (°C)
      required: [min_temp, max_temp, mean_temp]
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          type: "null"
        error:
          type: string
          description: Human-readable error description
          example: "GEE computation failed: User memory limit exceeded"
        cached:
          type: boolean
          example: false
      required: [success, error, cached]
  responses:
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            data: null
            error: "bbox area exceeds maximum of 50,000 km²"
            cached: false
    GEEError:
      description: GEE computation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            data: null
            error: "GEE service unavailable. Try again shortly."
            cached: false


# ─────────────────────────────────────────────────────────────────────────────
# ENDPOINTS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /api/heatmap:
    post:
      summary: Get LST heatmap tile URL
      description: |
        Returns a GEE tile URL for MODIS Land Surface Temperature (LST) over the
        specified bounding box and date range. The tile URL is consumed directly by
        Leaflet — pixel data does not pass through this API.

        Data source: MODIS MOD11A2, band LST_Day_1km
        Units: Celsius (scale factor and offset applied by backend)
        Typical response time: 3–15 seconds (GEE computation)
        Cache TTL: 6 hours
      operationId: get_heatmap
      tags: [Heatmap]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                bbox:
                  $ref: '#/components/schemas/BBox'
                start_date:
                  $ref: '#/components/schemas/DateString'
                end_date:
                  $ref: '#/components/schemas/DateString'
              required: [bbox, start_date, end_date]
            example:
              bbox: [72.77, 18.89, 72.99, 19.27]
              start_date: "2023-05-01"
              end_date: "2023-05-31"
      responses:
        "200":
          description: Heatmap tile URL and temperature statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      tile_url:
                        type: string
                        description: Leaflet-compatible GEE tile URL template ({x}, {y}, {z})
                        example: "https://earthengine.googleapis.com/v1/projects/.../maps/.../tiles/{z}/{x}/{y}"
                      stats:
                        $ref: '#/components/schemas/TemperatureStats'
                  error:
                    type: "null"
                  cached:
                    type: boolean
                    example: false
                required: [success, data, error, cached]
        "422":
          $ref: '#/components/responses/ValidationError'
        "503":
          $ref: '#/components/responses/GEEError'
...